// src/parser-langium/generated/ast.ts
export interface AdditiveExpr extends langium.AstNode {
    readonly $container: AdditiveExpr | AndExpr | AssignmentExpr | EqualityExpr | MultiplicationExpr | NamedParameter | NotExpr | OrExpr | ParExpr | Parameters | ParenthesizedExpr | RelationalExpr | Ret | UnaryMinusExpr;
    readonly $type: 'AdditiveExpr';
    left: Expr;
    op: string;
    right: Expr;
}

export const AdditiveExpr = {
    $type: 'AdditiveExpr',
    left: 'left',
    op: 'op',
    right: 'right'
} as const;

export function isAdditiveExpr(item: unknown): item is AdditiveExpr {
    return reflection.isInstance(item, AdditiveExpr.$type);
}

export interface Alt extends langium.AstNode {
    readonly $container: Block;
    readonly $type: 'Alt';
    elseBlock?: ElseBlock;
    elseIfBlocks: Array<ElseIfBlock>;
    ifBlock?: IfBlock;
}

export const Alt = {
    $type: 'Alt',
    elseBlock: 'elseBlock',
    elseIfBlocks: 'elseIfBlocks',
    ifBlock: 'ifBlock'
} as const;

export function isAlt(item: unknown): item is Alt {
    return reflection.isInstance(item, Alt.$type);
}

export interface AndExpr extends langium.AstNode {
    readonly $container: AdditiveExpr | AndExpr | AssignmentExpr | EqualityExpr | MultiplicationExpr | NamedParameter | NotExpr | OrExpr | ParExpr | Parameters | ParenthesizedExpr | RelationalExpr | Ret | UnaryMinusExpr;
    readonly $type: 'AndExpr';
    left: Expr;
    op: string;
    right: Expr;
}

export const AndExpr = {
    $type: 'AndExpr',
    left: 'left',
    op: 'op',
    right: 'right'
} as const;

export function isAndExpr(item: unknown): item is AndExpr {
    return reflection.isInstance(item, AndExpr.$type);
}

export interface Assignee extends langium.AstNode {
    readonly $container: Assignment;
    readonly $type: 'Assignee';
    atom?: Atom;
    ids: Array<string>;
}

export const Assignee = {
    $type: 'Assignee',
    atom: 'atom',
    ids: 'ids'
} as const;

export function isAssignee(item: unknown): item is Assignee {
    return reflection.isInstance(item, Assignee.$type);
}

export interface Assignment extends langium.AstNode {
    readonly $container: AssignmentExpr | CreationBody | MessageBody;
    readonly $type: 'Assignment';
    assignee?: Assignee;
    type?: Name;
}

export const Assignment = {
    $type: 'Assignment',
    assignee: 'assignee',
    type: 'type'
} as const;

export function isAssignment(item: unknown): item is Assignment {
    return reflection.isInstance(item, Assignment.$type);
}

export interface AssignmentExpr extends langium.AstNode {
    readonly $container: AdditiveExpr | AndExpr | AssignmentExpr | EqualityExpr | MultiplicationExpr | NamedParameter | NotExpr | OrExpr | ParExpr | Parameters | ParenthesizedExpr | RelationalExpr | Ret | UnaryMinusExpr;
    readonly $type: 'AssignmentExpr';
    assignment: Assignment;
    expr: Expr;
}

export const AssignmentExpr = {
    $type: 'AssignmentExpr',
    assignment: 'assignment',
    expr: 'expr'
} as const;

export function isAssignmentExpr(item: unknown): item is AssignmentExpr {
    return reflection.isInstance(item, AssignmentExpr.$type);
}

export interface AsyncMessage extends langium.AstNode {
    readonly $container: Block | Ret;
    readonly $type: 'AsyncMessage';
    content?: string;
    from?: From;
    to?: To;
}

export const AsyncMessage = {
    $type: 'AsyncMessage',
    content: 'content',
    from: 'from',
    to: 'to'
} as const;

export function isAsyncMessage(item: unknown): item is AsyncMessage {
    return reflection.isInstance(item, AsyncMessage.$type);
}

export type Atom = BooleanAtom | DigitLeadingNameAtom | IdAtom | MoneyAtom | NilAtom | NumberAtom | NumberUnitAtom | StringAtom;

export const Atom = {
    $type: 'Atom'
} as const;

export function isAtom(item: unknown): item is Atom {
    return reflection.isInstance(item, Atom.$type);
}

export interface AtomExpr extends langium.AstNode {
    readonly $container: AdditiveExpr | AndExpr | AssignmentExpr | EqualityExpr | MultiplicationExpr | NamedParameter | NotExpr | OrExpr | ParExpr | Parameters | ParenthesizedExpr | RelationalExpr | Ret | UnaryMinusExpr;
    readonly $type: 'AtomExpr';
    atom: Atom;
}

export const AtomExpr = {
    $type: 'AtomExpr',
    atom: 'atom'
} as const;

export function isAtomExpr(item: unknown): item is AtomExpr {
    return reflection.isInstance(item, AtomExpr.$type);
}

export interface Block extends langium.AstNode {
    readonly $container: BraceBlock | Prog;
    readonly $type: 'Block';
    stats: Array<Stat>;
}

export const Block = {
    $type: 'Block',
    stats: 'stats'
} as const;

export function isBlock(item: unknown): item is Block {
    return reflection.isInstance(item, Block.$type);
}

export interface BooleanAtom extends langium.AstNode {
    readonly $container: Assignee | AtomExpr | Section;
    readonly $type: 'BooleanAtom';
    value: string;
}

export const BooleanAtom = {
    $type: 'BooleanAtom',
    value: 'value'
} as const;

export function isBooleanAtom(item: unknown): item is BooleanAtom {
    return reflection.isInstance(item, BooleanAtom.$type);
}

export interface BraceBlock extends langium.AstNode {
    readonly $container: CatchBlock | Creation | Critical | ElseBlock | ElseIfBlock | FinallyBlock | IfBlock | Loop | Message | Opt | Par | Section | TryBlock;
    readonly $type: 'BraceBlock';
    block?: Block;
}

export const BraceBlock = {
    $type: 'BraceBlock',
    block: 'block'
} as const;

export function isBraceBlock(item: unknown): item is BraceBlock {
    return reflection.isInstance(item, BraceBlock.$type);
}

export interface CatchBlock extends langium.AstNode {
    readonly $container: Tcf;
    readonly $type: 'CatchBlock';
    braceBlock?: BraceBlock;
    invocation?: Invocation;
}

export const CatchBlock = {
    $type: 'CatchBlock',
    braceBlock: 'braceBlock',
    invocation: 'invocation'
} as const;

export function isCatchBlock(item: unknown): item is CatchBlock {
    return reflection.isInstance(item, CatchBlock.$type);
}

export type Condition = Expr | InExpr | TextExpr;

export const Condition = {
    $type: 'Condition'
} as const;

export function isCondition(item: unknown): item is Condition {
    return reflection.isInstance(item, Condition.$type);
}

export interface Creation extends langium.AstNode {
    readonly $container: Block | CreationExpr;
    readonly $type: 'Creation';
    body?: CreationBody;
    braceBlock?: BraceBlock;
}

export const Creation = {
    $type: 'Creation',
    body: 'body',
    braceBlock: 'braceBlock'
} as const;

export function isCreation(item: unknown): item is Creation {
    return reflection.isInstance(item, Creation.$type);
}

export interface CreationBody extends langium.AstNode {
    readonly $container: Creation;
    readonly $type: 'CreationBody';
    assignment?: Assignment;
    construct?: Name;
    parameters?: Parameters;
}

export const CreationBody = {
    $type: 'CreationBody',
    assignment: 'assignment',
    construct: 'construct',
    parameters: 'parameters'
} as const;

export function isCreationBody(item: unknown): item is CreationBody {
    return reflection.isInstance(item, CreationBody.$type);
}

export interface CreationExpr extends langium.AstNode {
    readonly $container: AdditiveExpr | AndExpr | AssignmentExpr | EqualityExpr | MultiplicationExpr | NamedParameter | NotExpr | OrExpr | ParExpr | Parameters | ParenthesizedExpr | RelationalExpr | Ret | UnaryMinusExpr;
    readonly $type: 'CreationExpr';
    creation: Creation;
}

export const CreationExpr = {
    $type: 'CreationExpr',
    creation: 'creation'
} as const;

export function isCreationExpr(item: unknown): item is CreationExpr {
    return reflection.isInstance(item, CreationExpr.$type);
}

export interface Critical extends langium.AstNode {
    readonly $container: Block;
    readonly $type: 'Critical';
    braceBlock?: BraceBlock;
    parExpr?: ParExpr;
}

export const Critical = {
    $type: 'Critical',
    braceBlock: 'braceBlock',
    parExpr: 'parExpr'
} as const;

export function isCritical(item: unknown): item is Critical {
    return reflection.isInstance(item, Critical.$type);
}

export interface Declaration extends langium.AstNode {
    readonly $container: Parameters;
    readonly $type: 'Declaration';
    name?: string;
    type?: Name;
}

export const Declaration = {
    $type: 'Declaration',
    name: 'name',
    type: 'type'
} as const;

export function isDeclaration(item: unknown): item is Declaration {
    return reflection.isInstance(item, Declaration.$type);
}

export interface DigitLeadingNameAtom extends langium.AstNode {
    readonly $container: Assignee | AtomExpr | Section;
    readonly $type: 'DigitLeadingNameAtom';
    value: string;
}

export const DigitLeadingNameAtom = {
    $type: 'DigitLeadingNameAtom',
    value: 'value'
} as const;

export function isDigitLeadingNameAtom(item: unknown): item is DigitLeadingNameAtom {
    return reflection.isInstance(item, DigitLeadingNameAtom.$type);
}

export interface Divider extends langium.AstNode {
    readonly $container: Block;
    readonly $type: 'Divider';
    note?: string;
}

export const Divider = {
    $type: 'Divider',
    note: 'note'
} as const;

export function isDivider(item: unknown): item is Divider {
    return reflection.isInstance(item, Divider.$type);
}

export interface ElseBlock extends langium.AstNode {
    readonly $container: Alt;
    readonly $type: 'ElseBlock';
    braceBlock?: BraceBlock;
}

export const ElseBlock = {
    $type: 'ElseBlock',
    braceBlock: 'braceBlock'
} as const;

export function isElseBlock(item: unknown): item is ElseBlock {
    return reflection.isInstance(item, ElseBlock.$type);
}

export interface ElseIfBlock extends langium.AstNode {
    readonly $container: Alt;
    readonly $type: 'ElseIfBlock';
    braceBlock?: BraceBlock;
    parExpr?: ParExpr;
}

export const ElseIfBlock = {
    $type: 'ElseIfBlock',
    braceBlock: 'braceBlock',
    parExpr: 'parExpr'
} as const;

export function isElseIfBlock(item: unknown): item is ElseIfBlock {
    return reflection.isInstance(item, ElseIfBlock.$type);
}

export interface Emoji extends langium.AstNode {
    readonly $container: From | MethodName | Participant | To;
    readonly $type: 'Emoji';
    name?: Name;
}

export const Emoji = {
    $type: 'Emoji',
    name: 'name'
} as const;

export function isEmoji(item: unknown): item is Emoji {
    return reflection.isInstance(item, Emoji.$type);
}

export interface EqualityExpr extends langium.AstNode {
    readonly $container: AdditiveExpr | AndExpr | AssignmentExpr | EqualityExpr | MultiplicationExpr | NamedParameter | NotExpr | OrExpr | ParExpr | Parameters | ParenthesizedExpr | RelationalExpr | Ret | UnaryMinusExpr;
    readonly $type: 'EqualityExpr';
    left: Expr;
    op: string;
    right: Expr;
}

export const EqualityExpr = {
    $type: 'EqualityExpr',
    left: 'left',
    op: 'op',
    right: 'right'
} as const;

export function isEqualityExpr(item: unknown): item is EqualityExpr {
    return reflection.isInstance(item, EqualityExpr.$type);
}

export type Expr = AdditiveExpr | AndExpr | AssignmentExpr | AtomExpr | CreationExpr | EqualityExpr | FuncExpr | MultiplicationExpr | NotExpr | OrExpr | ParenthesizedExpr | RelationalExpr | UnaryMinusExpr;

export const Expr = {
    $type: 'Expr'
} as const;

export function isExpr(item: unknown): item is Expr {
    return reflection.isInstance(item, Expr.$type);
}

export interface FinallyBlock extends langium.AstNode {
    readonly $container: Tcf;
    readonly $type: 'FinallyBlock';
    braceBlock?: BraceBlock;
}

export const FinallyBlock = {
    $type: 'FinallyBlock',
    braceBlock: 'braceBlock'
} as const;

export function isFinallyBlock(item: unknown): item is FinallyBlock {
    return reflection.isInstance(item, FinallyBlock.$type);
}

export interface From extends langium.AstNode {
    readonly $container: AsyncMessage | FromTo | ReturnAsyncMessage;
    readonly $type: 'From';
    emoji?: Emoji;
    name?: Name;
}

export const From = {
    $type: 'From',
    emoji: 'emoji',
    name: 'name'
} as const;
// ...
export interface Group extends langium.AstNode {
    readonly $container: Prog;
    readonly $type: 'Group';
    name?: Name;
    participants: Array<Participant>;
}

export const Group = {
    $type: 'Group',
    name: 'name',
    participants: 'participants'
} as const;
// ...
export interface Participant extends langium.AstNode {
    readonly $container: Group | Prog;
    readonly $type: 'Participant';
    color?: string;
    emoji?: Emoji;
    label?: Label;
    name?: Name;
    participantType?: string;
    stereotype?: Stereotype;
    width?: string;
}

export const Participant = {
    $type: 'Participant',
    color: 'color',
    emoji: 'emoji',
    label: 'label',
    name: 'name',
    participantType: 'participantType',
    stereotype: 'stereotype',
    width: 'width'
} as const;

export function isParticipant(item: unknown): item is Participant {
    return reflection.isInstance(item, Participant.$type);
}

export interface Prog extends langium.AstNode {
    readonly $type: 'Prog';
    block?: Block;
    headElements: Array<Group | Participant>;
    starterExp?: StarterExp;
    title?: Title;
}

export const Prog = {
    $type: 'Prog',
    block: 'block',
    headElements: 'headElements',
    starterExp: 'starterExp',
    title: 'title'
} as const;
// ...
export interface StarterExp extends langium.AstNode {
    readonly $container: Prog;
    readonly $type: 'StarterExp';
    annotation?: string;
    starter?: Name;
}

export const StarterExp = {
    $type: 'StarterExp',
    annotation: 'annotation',
    starter: 'starter'
} as const;
// ...
export interface Title extends langium.AstNode {
    readonly $container: Prog;
    readonly $type: 'Title';
    content?: string;
}

export const Title = {
    $type: 'Title',
    content: 'content'
} as const;
// ...
export type ZenUmlAstType = {
    AdditiveExpr: AdditiveExpr
    Alt: Alt
    AndExpr: AndExpr
    Assignee: Assignee
    Assignment: Assignment
    AssignmentExpr: AssignmentExpr
    AsyncMessage: AsyncMessage
    Atom: Atom
    AtomExpr: AtomExpr
    Block: Block
    BooleanAtom: BooleanAtom
    BraceBlock: BraceBlock
    CatchBlock: CatchBlock
    Condition: Condition
    Creation: Creation
    CreationBody: CreationBody
    CreationExpr: CreationExpr
    Critical: Critical
    Declaration: Declaration
    DigitLeadingNameAtom: DigitLeadingNameAtom
    Divider: Divider
    ElseBlock: ElseBlock
    ElseIfBlock: ElseIfBlock
    Emoji: Emoji
    EqualityExpr: EqualityExpr
    Expr: Expr
    FinallyBlock: FinallyBlock
    From: From
    FromTo: FromTo
    Func: Func
    FuncExpr: FuncExpr
    Group: Group
    IdAtom: IdAtom
    IfBlock: IfBlock
    InExpr: InExpr
    Invocation: Invocation
    Label: Label
    Loop: Loop
    Message: Message
    MessageBody: MessageBody
    MethodName: MethodName
    MoneyAtom: MoneyAtom
    MultiplicationExpr: MultiplicationExpr
    NamedParameter: NamedParameter
    NilAtom: NilAtom
    NotExpr: NotExpr
    NumberAtom: NumberAtom
    NumberUnitAtom: NumberUnitAtom
    Opt: Opt
    OrExpr: OrExpr
    Par: Par
    ParExpr: ParExpr
    Parameter: Parameter
    Parameters: Parameters
    ParenthesizedExpr: ParenthesizedExpr
    Participant: Participant
    Prog: Prog
    Ref: Ref
    RelationalExpr: RelationalExpr
    Ret: Ret
    ReturnAsyncMessage: ReturnAsyncMessage
    Section: Section
    Signature: Signature
    StarterExp: StarterExp
    Stat: Stat
    Stereotype: Stereotype
    StringAtom: StringAtom
    Tcf: Tcf
    TextExpr: TextExpr
    Title: Title
    To: To
    TryBlock: TryBlock
    UnaryMinusExpr: UnaryMinusExpr
}